chore: :truck: move code to independent folders
