+++ src/app/components/select/option/ui-select-option.component.scss
