+++ src/app/components/select/panel/ui-select-panel.component.scss
