+++ src/app/components/select/panel/ui-select-panel.component.ts
@@ -7,12 +7,12 @@ import {
     Input,
     QueryList,
     ViewChild,
-    ViewEncapsulation,
+    ViewEncapsulation
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 import { fadeInOut } from '@components/animations/fade.animation';
-import { UiSelectOptionComponent } from './ui-select-option.component';
-import { UiSelectService } from './ui-select.service';
+import { UiSelectOptionComponent } from '../option/ui-select-option.component';
+import { UiSelectService } from '../ui-select.service';
 
 @Component({
     // eslint-disable-next-line @angular-eslint/component-selector
