+++ src/app/components/select/ui-select.module.ts
@@ -1,8 +1,8 @@
 import { OverlayModule } from '@angular/cdk/overlay';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { UiSelectOptionComponent } from './ui-select-option.component';
-import { UiSelectPanelComponent } from './ui-select-panel.component';
+import { UiSelectOptionComponent } from './option/ui-select-option.component';
+import { UiSelectPanelComponent } from './panel/ui-select-panel.component';
 
 @NgModule({
     declarations: [UiSelectPanelComponent, UiSelectOptionComponent],
